+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.3.0 - ??
+
+### Gameplay
+
+* Reduced the cost of several tier 2 and 3 lights
+
+### Daily Challenges
+
+* "Normal" encounters no longer duplicate, and "hard" encounters don't duplicate the normal at the same depth
+* If the player can't afford 3 items, the shop will fill with weak $1 items
+
 ## 1.2.0 - 3/19/2026
 
 ### Daily Challenges